+++ src/core/darwin/fruitjector-glue.c
@@ -62,7 +62,7 @@ struct _ZedInjectionInstance
 
 struct _ZedAgentDetails
 {
-  gulong pid;
+  guint pid;
   const char * dylib_path;
   const char * data_string;
   GumCpuType cpu_type;
@@ -211,7 +211,7 @@ _zed_fruitjector_free_instance (ZedFruitjector * self, void * instance)
 }
 
 guint
-_zed_fruitjector_do_inject (ZedFruitjector * self, gulong pid, const gchar * dylib_path, const char * data_string, GError ** error)
+_zed_fruitjector_do_inject (ZedFruitjector * self, guint pid, const gchar * dylib_path, const char * data_string, GError ** error)
 {
   ZedFruitContext * ctx = self->context;
   ZedInjectionInstance * instance;
